test: move shared theme helper into support file
