+++ site/tests/e2e/component_interaction_contrast_test.go
@@ -19,26 +19,6 @@ type renderedContrast struct {
 	Filter     string  `json:"filter"`
 }
 
-func setThemeMode(t *testing.T, page playwright.Page, theme string, dark bool) {
-	t.Helper()
-	_, err := page.Evaluate(`([theme, dark]) => {
-		const html = document.documentElement;
-		let style = document.getElementById('contrast-test-no-transitions');
-		if (!style) {
-			style = document.createElement('style');
-			style.id = 'contrast-test-no-transitions';
-			style.textContent = '* { transition: none !important; }';
-			document.head.append(style);
-		}
-		localStorage.setItem('theme', theme);
-		html.setAttribute('data-theme', theme);
-		html.classList.toggle('dark', dark);
-	}`, []any{theme, dark})
-	require.NoError(t, err)
-	_, err = page.WaitForFunction("theme => document.documentElement.dataset.theme === theme", theme)
-	require.NoError(t, err)
-}
-
 func measureRenderedContrast(t *testing.T, locator playwright.Locator) renderedContrast {
 	return measureRenderedPropertyContrast(t, locator, "color")
 }
